+++ myUtils/nodeEncode/base64_IntArray.go
@@ -1,9 +1,9 @@
 package nodeEncode
 
-func IntArrayToBase64Str(intArray []int) string {
+func IntArrayToBase64Str(intArray []int64) string {
 	return ToBase64(IntArrayToByteArray(intArray))
 }
-func IntArrayToByteArray(intArray []int) []byte {
+func IntArrayToByteArray(intArray []int64) []byte {
 	byteArray := make([]byte, max(intArray)/8+1)
 	for _, v := range intArray {
 		quotient := v / 8
@@ -13,7 +13,7 @@ func IntArrayToByteArray(intArray []int) []byte {
 	return byteArray
 }
 
-func Base64StrToIntArray(base64Str string) ([]int, error) {
+func Base64StrToIntArray(base64Str string) ([]int64, error) {
 	byteArray, err := ToByteArray(base64Str)
 	if err != nil {
 		return nil, err
@@ -21,19 +21,19 @@ func Base64StrToIntArray(base64Str string) ([]int, error) {
 	return ByteArrayToIntArray(byteArray), nil
 }
 
-func ByteArrayToIntArray(byteArray []byte) []int {
-	intArray := make([]int, 0)
+func ByteArrayToIntArray(byteArray []byte) []int64 {
+	intArray := make([]int64, 0)
 	for i := 0; i < len(byteArray); i++ {
 		for j := 0; j < 8; j++ {
 			if byteArray[i]&(1<<j) != 0 {
-				intArray = append(intArray, j+8*i)
+				intArray = append(intArray, int64(j+8*i))
 			}
 		}
 	}
 	return intArray
 }
 
-func max(theArray []int) (theMax int) {
+func max(theArray []int64) (theMax int64) {
 	theMax = theArray[0]
 	for _, v := range theArray {
 		if v > theMax {
